Rename API routes to /registration-requests/
